Unroll `Upload` deletion and run it in chunks
The recently introduces code to delete old `Upload`s does not function properly.
The job seems to hang for hours at a time, and does not seem to do any meaningful progress.

I’m unsure what the reason for that is, but I think the solution to this is to make the `Upload` deletion work in chunks. We are still using the time-range constrained queries to fetch `Upload` ids, but then do the deletion based on those fetched ids, which should split up that problem into smaller chunks, so those can then make progress.
